card 속성 private(set)으로 변경, Playable.swift 생성
 - 플레이어 등록 시 dealer 매개변수를 Playable 수정
 - Playable안에서 카드 점수 구현할 수 있는 extension 구현,
